Expand to 7 states in qemu test
The default "max_hangs" is 5, so we want to look at 7 total states:
 * the initial state0 is never considered hung
 * state1 through state5 all increment the hang_count
   * but hung is when the hang count _exceeds_ max_hangs
 * state6 will increment the hang_count to greater than 5
   * this now truly looks hung

Re-imported the list of qemu proc data to include two additional states.

Now the test can use the default hang_check and _not_ need hangs_for_len().

It seems we altered the data prior, but not clear as to why, so I didn't
bother doing that again, though the diff looks bigger than it should.